Downloads/withdrawbot-bugfix-message-edit/withdrawbot-bugfix-message-edit/requirements.txt
aiohttp
python-telegram-bot
python-dotenv
flask
gunicorn
aiosqlite
requests
Werkzeug
Jinja2
click
itsdangerous
MarkupSafe
logger
psycopg2-binary
pandas
xlsxwriter
base58
openpyxl
pytz
